Tout est fixé concernant les calcul et l'impression sur html et printtotex _FINALCOMMIT

+++ src/main/java/Customer.java
@@ -1,10 +1,10 @@
 public class Customer {
     String name;
     int numcust;
-    double soldepoints;
+    int soldepoints;
 
 
-    public Customer(String name, int numcust, double soldepoints) {
+    public Customer(String name, int numcust, int soldepoints) {
         this.name = name;
         this.numcust = numcust;
         this.soldepoints = soldepoints;
@@ -16,10 +16,10 @@ public class Customer {
       
       return numcust;
     }
-    public double getsoldepoints(){
+    public int getsoldepoints(){
       return soldepoints;
     }
-    public void setsoldepoints(double soldepoints){
+    public void setsoldepoints(int soldepoints){
       this.soldepoints=soldepoints;
     }
 }
